portability fix: pointer-to-int conversion

+++ storm/storm3dv2/storm3d_terrain_utils.h
@@ -176,7 +176,7 @@ public:
 	void unlock();
 
 	void apply(int stream) const;
-	void apply(int stream, int offset) const;
+	void apply(int stream, uintptr_t offset) const;
 	operator bool() const;
 };
